// packages/resources-manager/src/browser/resources/types.ts
export type ConfigResourceValues = {
  name: string;
  id: string;
  localPath: string;
  remoteUrl: string;
  version: string;
  type: string;
  [x: string]: any;
};